♻️(models) move some models to separate files
As the Referral model keeps growing, we want to move everything
that makes sense to other files in our models folder so as to make
it easier to work with them.
