+++ src/backend/partaj/core/models/__init__.py
@@ -6,5 +6,7 @@ Make models.py a module containing one file per model as it was getting too long
 from .attachment import *
 from .referral import *
 from .referral_activity import *
+from .referral_answer import *
 from .referral_message import *
+from .referral_urgency import *
 from .unit import *
